Simplify ScopedString interphase and parsing improvements
Firstly this should clear up some confusion about scopes. In this
project scope means some form of lexical scope (basically what is
between the begin and end pattern) while what TextMate calles scope
selectors or scope for short is now called patternIdentifier or
identifier for short. A scope is now also a type of result.
Added test for the range extensions.
The parser now returns the range that definitely has to	be parsed but
parses further if it has to. Turns out the range cannot	be determined
without actually parsing the new string so this seems to be a good
tradeoff.